# Conflicts:
#	.env
#	core/model.php
#	models/model_Add_AdminSite.php
#	next.config.js

+++ models/model_Add_AdminSite.php
@@ -2,7 +2,7 @@
 
 class model_Add_AdminSite extends Model
 {
-    public $url = 'http://localhost:3000';
+    public $url = 'http://localhost';
 
     function __construct()
     {
@@ -36,7 +36,7 @@ class model_Add_AdminSite extends Model
         $result = $this->doSelect($sql, []);
         foreach ($result as $key => $item) {
             $tech = $item['technology'];
-            $sql2 = 'select * from technology where code IN (' . $tech . ')  ';
+            $sql2 = 'select * from technology where code IN (' . $tech . ') ';
             $resultTech = $this->doSelect($sql2);
             $result[$key]['technology'] = $resultTech;
         }
@@ -141,10 +141,13 @@ class model_Add_AdminSite extends Model
             $nameProject = $this->validationFrom($data['nameProject']);
             $linkProject = $this->validationFrom($data['linkProject']);
             $technology = join(',', $data['technology']);
+
             if (empty($id)) {
+
                 $sql = 'insert into projects (title,link,technology) values (?,?,?)';
                 $this->doQuery($sql, [$nameProject, $linkProject, $technology]);
                 $lastId = self::$conn->lastInsertId();
+
                 if (isset($file['imageProject'])) {
 
                     mkdir('public/image/image_project/' . $lastId . '/');
@@ -184,62 +187,64 @@ class model_Add_AdminSite extends Model
                         $this->doQuery($sql, [$nameImageTable, $target, $lastId]);
 
                     }
+
                 }
+
                 header('location:' . $this->url . '/dashboardamircv');
 
             } else {
-                
-                $sql = 'update projects set title=?,link=?,technology=? where id=?';
-                $this->doQuery($sql, [$nameProject, $linkProject, $technology, $id]);
-                if (isset($file['imageProject'])) {
 
-                    $fileAttr = $file['imageProject'];
-                    $imageName = $fileAttr['name'];
-                    $fileTyp = $fileAttr['type'];
-                    $fileTmp_name = $fileAttr['tmp_name'];
-                    $fileError = $fileAttr['error'];
-                    $fileSize = $fileAttr['size'];
-                    $uploadYes = 1;
-                    $insertYes = 1;
-                    $randomTime = time() * 2;
-                    $newName = $randomTime;
-                    $targetMine = 'public/image/image_project/';
-                    $ImageSave_Nwe_Dir = $targetMine . $id . '/';
-                    if ($fileTyp !== 'image/jpeg' and $fileTyp !== 'image/jpg' and $fileTyp !== 'image/png' and $fileTyp !== 'image/svg+xml') {
-                        $uploadYes = 0;
-                        $insertYes = 0;
+                if (!empty($technology)){
+                    $sql = 'update projects set title=?,link=?,technology=? where id=?';
+                    $this->doQuery($sql, [$nameProject, $linkProject, $technology, $id]);
+                    if (isset($file['imageProject'])) {
 
-                    }
-                    if ($fileSize <= 0) {
+                        $fileAttr = $file['imageProject'];
+                        $imageName = $fileAttr['name'];
+                        $fileTyp = $fileAttr['type'];
+                        $fileTmp_name = $fileAttr['tmp_name'];
+                        $fileError = $fileAttr['error'];
+                        $fileSize = $fileAttr['size'];
+                        $uploadYes = 1;
+                        $insertYes = 1;
+                        $randomTime = time() * 2;
+                        $newName = $randomTime;
+                        $targetMine = 'public/image/image_project/';
+                        $ImageSave_Nwe_Dir = $targetMine . $id . '/';
+                        if ($fileTyp !== 'image/jpeg' and $fileTyp !== 'image/jpg' and $fileTyp !== 'image/png' and $fileTyp !== 'image/svg+xml') {
+                            $uploadYes = 0;
+                            $insertYes = 0;
 
-                        $uploadYes = 0;
-                        $insertYes = 0;
-                    }
-                    if ($fileSize > 41453494304) {
-                        $uploadYes = 0;
-                        $insertYes = 0;
-                    }
-                    if ($uploadYes == 1) {
-
-                        $selectImage = 'select tmp_image from projects where id=?';
-                        $result = $this->doSelect($selectImage, [$id], 1);
-                        if (file_exists($result['tmp_image'])) {
-                            unlink($result['tmp_image']);
                         }
-                        $exe = pathinfo($imageName, PATHINFO_EXTENSION);
-                        $target = $ImageSave_Nwe_Dir . $newName . '.' . $exe;
-                        $nameImageTable = '/image/image_project/' . $id . '/' . $newName . '.' . $exe;
-                        move_uploaded_file($fileTmp_name, $target);
-                        $sql = 'update projects set image=?,tmp_image=? where id=?';
-                        $this->doQuery($sql, [$nameImageTable, $target, $id]);
+                        if ($fileSize <= 0) {
+
+                            $uploadYes = 0;
+                            $insertYes = 0;
+                        }
+                        if ($fileSize > 41453494304) {
+                            $uploadYes = 0;
+                            $insertYes = 0;
+                        }
+                        if ($uploadYes == 1) {
+
+                            $selectImage = 'select tmp_image from projects where id=?';
+                            $result = $this->doSelect($selectImage, [$id], 1);
+                            if (file_exists($result['tmp_image'])) {
+                                unlink($result['tmp_image']);
+                            }
+                            $exe = pathinfo($imageName, PATHINFO_EXTENSION);
+                            $target = $ImageSave_Nwe_Dir . $newName . '.' . $exe;
+                            $nameImageTable = '/image/image_project/' . $id . '/' . $newName . '.' . $exe;
+                            move_uploaded_file($fileTmp_name, $target);
+                            $sql = 'update projects set image=?,tmp_image=? where id=?';
+                            $this->doQuery($sql, [$nameImageTable, $target, $id]);
+
+                        }
 
                     }
-
                 }
-                header('location:' . $this->url . '/dashboardamircv');
+                header('location:' . $this->url . '/dashboardamircv/');
             }
-
-
 
         }
 
@@ -381,5 +386,11 @@ class model_Add_AdminSite extends Model
         return json_encode($admin);
     }
 
+    function editSelect($data){
+        $dataSelect = $data['ids'];
+        $sql2 = 'select label,value,image from technology where code IN (' . $dataSelect . ') ';
+        $resultTech = $this->doSelect($sql2);
+        return json_encode($resultTech);
+    }
 }
